Move Waybar status JSON contract into the library
`format_state_json`, `ExtendedStatusInfo`, and `backend_display_for_variant`
moved from `src/app/status.rs` (binary-only) into a new `src/status_json.rs`
library module. The architecture review flagged this as a Rule-5 risk: the
JSON shape is the external API voxtype exposes to every status consumer
(Waybar via `voxtype status --follow`, Quickshell, the planned audio bridge,
a hypothetical future StatusNotifierItem tray), and burying it behind the
binary boundary meant any future caller would have to either re-implement
the shape or shell out to the CLI.

What stays in `src/app/status.rs`: only `run_status`, the inotify follow
loop and state-file IO. That's correctly binary-side — it's `voxtype
status` the subcommand, not the status JSON contract itself.

The three pin tests (`record_cancel_leaves_idle_not_stopped`,
`format_state_json_pins_byte_exact_output`,
`format_state_json_escapes_quotes_and_backslashes`) move with the function
they pin. Test count: 774 lib + 1 bin = 775 total, same as before.

Behaviour-preserving: clippy clean, fmt clean, all 9 subcommand --help
outputs byte-identical to the dev baseline. The new module is `pub` from
`voxtype::status_json`, accessible to any future in-process consumer.
